Update red/blue button style for item preview
It works for buttons present when previewing cosmetics and war paints.

`PaintkitPreviewContainer` was shifted downwards so the
buttons do not overlap it.

I don't have a reliable way to test changes in
collectioncraftingdialog_base.res and importpreviewitempanel.res so I've
done my best to guess how the buttons should look.